Edit on details screen

+++ app/src/main/java/com/sgut/android/nationalfootballleague/domain/domainmodels/new_game_details/PickcenterModel.kt
@@ -3,8 +3,8 @@ package com.sgut.android.nationalfootballleague.domain.domainmodels.new_game_det
 data class PickcenterModel(
     val provider: ProviderModel = ProviderModel(),
     val details: String = "",
-    val overUnder: Float = 0f,
-    val spread: Float = 0f,
+    val overUnder: Double = 0.0,
+    val spread: Double = 0.0,
     val awayTeamOdds: AwayTeamOddsModel = AwayTeamOddsModel(),
     val homeTeamOdds: HomeTeamOddsModel = HomeTeamOddsModel(),
